use custom find_exporter
