+++ app/exporters/decidim/anonymous_codes/exporters/anonymous_tokens_pdf_controller_helper.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Decidim
+  module AnonymousCodes
+    module Exporters
+      # rubocop: disable Rails/ApplicationController
+      # A dummy controller to render views while exporting questionnaires
+      class AnonymousTokensPdfControllerHelper < ActionController::Base
+        # rubocop: enable Rails/ApplicationController
+        helper Decidim::TranslationsHelper
+      end
+    end
+  end
+end
