--- app/exporters/decidim/exporters/anonymous_tokens_pdf_controller_helper.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module Exporters
-    # rubocop: disable Rails/ApplicationController
-    # A dummy controller to render views while exporting questionnaires
-    class AnonymousTokensPdfControllerHelper < ActionController::Base
-      # rubocop: enable Rails/ApplicationController
-      helper Decidim::TranslationsHelper
-    end
-  end
-end
