+++ app/jobs/decidim/anonymous_codes/export_group_tokens_job.rb
@@ -6,7 +6,15 @@ module Decidim
       queue_as :exports
 
       def perform(user, group, format)
-        export_data = Decidim::Exporters.find_exporter(format).new(group.tokens, TokenSerializer).export
+        exporter = format == "AnonymousTokensPdf" ? Decidim::AnonymousCodes::Exporters::AnonymousTokensPdf : Decidim::Exporters.find_exporter(format)
+
+        if exporter
+          Rails.logger.info "Exporting tokens for group #{group.id} in #{format} format"
+        else
+          Rails.logger.error "Cannot export tokens for group #{group.id}: Unknown format: #{format}"
+        end
+
+        export_data = exporter.new(group.tokens, TokenSerializer).export
 
         ExportMailer.export(user, "tokens_for_group_#{group.id}", export_data).deliver_now
       end
